refactor: migrate MultiDb extension
 Migrate the MultiDb extensions, annotations, and helper classes
 to the util module we have for our integration tests
