+++ qa/integration-tests/src/test/java/io/camunda/it/client/CreateDocumentLinkTest.java
@@ -13,7 +13,7 @@ import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
 import io.camunda.client.CamundaClient;
 import io.camunda.client.api.command.ProblemException;
 import io.camunda.client.api.response.DocumentReferenceResponse;
-import io.camunda.it.utils.MultiDbTest;
+import io.camunda.qa.util.multidb.MultiDbTest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
