+++ qa/integration-tests/src/test/java/io/camunda/it/orchestration/ProcessMigrationIT.java
@@ -25,7 +25,7 @@ import io.camunda.client.api.search.response.ProcessDefinition;
 import io.camunda.client.api.search.response.ProcessInstance;
 import io.camunda.client.api.search.response.UserTask;
 import io.camunda.client.api.search.response.Variable;
-import io.camunda.it.utils.MultiDbTest;
+import io.camunda.qa.util.multidb.MultiDbTest;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
