+++ qa/integration-tests/src/test/java/io/camunda/it/utils/BrokerITInvocationProvider.java
@@ -16,6 +16,7 @@ import io.camunda.authentication.config.AuthenticationProperties;
 import io.camunda.client.CamundaClient;
 import io.camunda.qa.util.auth.Authenticated;
 import io.camunda.qa.util.auth.User;
+import io.camunda.qa.util.multidb.CamundaClientTestFactory;
 import io.camunda.security.configuration.ConfiguredUser;
 import io.camunda.security.configuration.InitializationConfiguration;
 import io.camunda.security.entity.AuthenticationMethod;
