+++ qa/util/src/main/java/io/camunda/qa/util/multidb/MultiDbTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.it.utils;
+package io.camunda.qa.util.multidb;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
